move get function into own file
