+++ chessdotcom/endpoints/__init__.py
@@ -5,6 +5,7 @@ from ..client import Client, Resource
 from ..response_builder import ChessDotComResponse
 from ..utils import resolve_date
 from .player_profile import get_player_profile
+from .player_stats import get_player_stats
 
 
 @Client.endpoint
@@ -18,22 +19,6 @@ def get_titled_players(
     """
     return Resource(
         uri=f"/titled/{title_abbrev}", tts=tts, request_options=request_options
-    )
-
-
-@Client.endpoint
-def get_player_stats(username: str, tts=0, **request_options) -> ChessDotComResponse:
-    """
-    :param username: username of the player.
-    :param tts: the time the client will wait before making the first request.
-    :returns: ``ChessDotComResponse`` object containing information about the
-                plyers's ratings, win/loss, and other stats.
-    """
-    return Resource(
-        uri=f"/player/{username}/stats",
-        tts=tts,
-        top_level_attribute="stats",
-        request_options=request_options,
     )
